All SWTBot tests in a single file
