--- edelta.parent/edelta.swtbot.tests/src/edelta/swtbot/tests/EdeltaTemplateGeneratorSwtBotTest.java
@@ -1,39 +1,0 @@
-package edelta.swtbot.tests;
-
-import static org.eclipse.xtext.ui.testing.util.IResourcesSetupUtil.waitForBuild;
-
-import org.eclipse.core.runtime.CoreException;
-import org.eclipse.swtbot.swt.finder.junit.SWTBotJunit4ClassRunner;
-import org.junit.BeforeClass;
-import org.junit.Test;
-import org.junit.runner.RunWith;
-
-import edelta.swtbot.tests.utils.ProjectImportUtil;
-
-/**
- * @author Lorenzo Bettini
- *
- */
-@RunWith(SWTBotJunit4ClassRunner.class)
-public class EdeltaTemplateGeneratorSwtBotTest extends EdeltaAbstractSwtbotTest {
-
-	private static final String TEST_PROJECT = "edelta.dependency.analyzer.swtbot.tests.project";
-	private static final String PERSONS_MM_ECORE = "PersonsMM.ecore";
-	private static final String PERSONS_MM_TEMPLATE_EDELTA = "PersonsMMTemplate.edelta";
-
-	@BeforeClass
-	public static void importTestProject() throws Exception {
-		ProjectImportUtil.importJavaProject(TEST_PROJECT);
-		waitForBuild();
-	}
-
-	@Test
-	public void canGenerateTemplateFiles() throws CoreException {
-		assertErrorsInProject(0);
-		edeltaContextMenu(TEST_PROJECT, PERSONS_MM_ECORE, "Generate Edelta Template File");
-		waitForBuild();
-		assertErrorsInProject(0);
-		bot.editorByTitle(PERSONS_MM_TEMPLATE_EDELTA);
-	}
-
-}
